tests: Add a simple test of -o local -of qcow2 -oo compressed
This only tests that it doesn't completely fail, which it did before
we fixed nbdcopy.  I checked the file sizes manually and with
compression the resulting file is about half the size.

This test is a clone of tests/test-v2v-of-option.sh.  In order to
compare the sizes across the two tests, and to keep the tests fairly
similar I added an ls -l command to the original test.

Reviewed-by: Laszlo Ersek <lersek@redhat.com>