convert_linux: remove LVM2 "devices file"
A recent feature of LVM2 is the "devices file"
<https://man7.org/linux/man-pages/man8/lvmdevices.8.html>. It speeds up
LVM2 PV discovery on a normal system, but an old devices file in a
converted domain (with different hardware) can prevent the assembly of
volume groups.

In particular, when converting a physical system to a guest with virt-p2v,
the original system will have used "sys_wwid"-type identifiers in the LVM2
devices file, and those are guaranteed not to match any virtio-blk disks
in the output domain.

We've seen a similar issue in the past under RHBZ#1164853, so just extend
the same scrubbing with the new pathname.

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=2112801
Signed-off-by: Laszlo Ersek <lersek@redhat.com>
Message-Id: <20220805084426.9200-1-lersek@redhat.com>
Acked-by: Richard W.M. Jones <rjones@redhat.com>