rhv: Use osinfo to distinguish Windows >= 10 variants
Windows versions >= 10 no longer use the NT major.minor numbering
scheme (it is fixed at 10.0).  Libguestfs >= 1.49.8 can distinguish
these versions and it sets <osinfo> correctly, so use that instead.

After this change the OVF will contain:

  <Info>Windows 10 Pro</Info>
  <Description>windows_11</Description>

which is strange, but apparently what Microsoft intended.  As far as
RHV is concerned it only needs <Description> to choose the correct
devices etc.

Reported-by: Tingting Zheng
Fixes: https://bugzilla.redhat.com/show_bug.cgi?id=2149863
Reviewed-by: Laszlo Ersek <lersek@redhat.com>