convert: Add a handle type to Windows_virtio
To further encapsulate the Windows_virtio module, add a handle.
Initially there is just one way to set up the handle, through
Windows_virtio.from_environment which reads the $VIRTIO_WIN
environment variable.

In future, when we move this module into common/mlcustomize, we will
add other ways to create this handle from a path.  (Using environment
variables with virt-customize is undesirable.)

I made further notes in comments about aspects of this that I think
are a real hack, but that can be addressed in future now that the
module is completely self-contained.

This is just refactoring and shouldn't change functionality at all.