convert: windows: Move code for injecting virtio-win to common/mlcustomize
So that we can reuse this useful code in virt-customize, move it to
the common module.

This is mostly code motion, but note:

(a) the module name changes from Windows_virtio to Inject_virtio_win

(b) as there were still some dependencies on types in the virt-v2v
Types module (eg. block driver type) I had to duplicate those.

See also:
https://listman.redhat.com/archives/libguestfs/2022-December/030389.html