Remove support for RHEL 3
This was the only guest type which did not support ACPI.  It also only
supported direct to libvirt output.  It hasn't been tested for a long
time, and hasn't been supported by Red Hat for very much longer.
Removing this means we no longer have to think about non-ACPI guests.