convert: Remove dependency of Linux module on Types
As this module is generally useful we would like to use the same code
in guestfs-tools (ie. move it into common/).  In preparation for doing
that, remove dependencies on the virt-v2v specific Types module.

This change is mostly routine avoidance of the inspect struct, but I
changed the way that we determine which version of RPM is installed.
Instead of consulting inspect.i_apps, actually run 'rpm --version'
(inside the guest) and parse the output.  Note the only place we do
this, we're going to run 'rpm' soon anyway, so we know this is OK.