-o libvirt: Use cpu='host-model' for gcaps_default_cpu = false
For RHEL >= 9 / x86-64 guests we cannot use the default qemu CPU
(eg. "qemu64"), and so we have a mechanism for conversion to indicate
to the output modes that a more capable CPU is required.  We
previously picked cpu='host-passthrough' (ie. the equivalent of qemu's
-cpu host).  However this is not live migratable.  cpu='host-model' is
a better choice as it is more likely to be migratable.

See also discussion here:
https://listman.redhat.com/archives/libguestfs/2023-February/030625.html

and previous discussion here:
https://listman.redhat.com/archives/libguestfs/2022-April/thread.html#28711

Acked-by: Laszlo Ersek <lersek@redhat.com>