tests: Fix network tests for newer libvirt
The old tests' network bridge had a bogus <mac address="00:00:00:00:00:00">.

Older libvirt appeared to drop this completely.  Newer libvirt appears
to rewrite it as a valid MAC, eg. <mac address="52:54:00:xx:xx:xx">
(for some randomly chosen "xx"'s).  To make the tests stable, replace
the old bogus value with a valid value, and update the tests to check
for it.