virt-v2v: -i vmx: Remove scp -T option
This reverts the following commit:

  commit d265639c2ab31418cfdbdedd0cc3e68cf290d834
  Author: Richard W.M. Jones <rjones@redhat.com>
  Date:   Thu Jul 25 14:52:42 2019 +0100

    v2v: -i vmx: Use scp -T option if available to unbreak scp (RHBZ#1733168).

See also the referenced bug:
https://bugzilla.redhat.com/show_bug.cgi?id=1733168

My rationale for removing this option is that since we now require
OpenSSH 8.8 we must be using sftp for file transfer so we no longer
need to defeat the check for correct expansion of wildcards.  That
check was only relevant for OpenSSH <= 8.7 using the old scp protocol.

Reverts: commit d265639c2ab31418cfdbdedd0cc3e68cf290d834