input/nbdkit_ssh: Make retry filter optional
This filter retries a request on failure.  Add a flag to make this
filter optional, but default to using it if available (so there's no
change to default behaviour).

The reason we want to make this optional is that the next commit will
want to use nbdkit-ssh-plugin to probe to see if a remote file is
present, and we don't want to retry that probe on failure.

Reviewed-by: Laszlo Ersek <lersek@redhat.com>