-o qemu: Remove final use of error_unless_uefi_firmware
This function was used by -o libvirt / -o local to check that we have
UEFI firmware for the target architecture.  All it does is called
find_uefi_firmware early so that we fail early.

The same check was copied into -o qemu, probably wrongly as we call
find_uefi_firmware just below, so the early fail is not necessary.

This allows us to remove error_unless_uefi_firmware completely since
it is no longer used anywhere.