tests/test-v2v-o-kubevirt-windows.sh: Set fake virtio-win
When converting this phony Windows guest we didn't set the VIRTIO_WIN
environment variable.  As a result the test output could change (and
break) depending on if the virtio-win package was installed or not.

Point VIRTIO_WIN to the fake virtio-win directory (as we do with other
Windows tests).  We also need to modify the test output since the fake
virtio-win does contain enough for virt-v2v to think that virtio
driver installation was successful.

Fixes: commit af1f7b571fb21fdbfa2974ee54a22b4fdcbc886a