-o rhv-upload: Output outside of the with block
We write the json inside a with block, so what actually happen is:

1. Computing the values
2. Dumping the json to stdout
3. Closing the connection

If some code in sdk.Connection.close() write something to stdout, it
will corrupt our json output. Avoid this by moving the output out of the
with block.

This will not prevent bad code to write something before the json. We
need to log bad json in the Ocaml code to reveal such issue.