-o rhv-upload: Improve host id logging
This log[1] is proved confusing for developers:

    cannot read /etc/vdsm/vdsm.id, using any host: [Errno 2] No such file or directory: '/etc/vdsm/vdsm.id'

This is actually a debug message, that can help RHV developers to
understand where virt-v2v was running. To make this more clear, in the
expected case when running on non-oVirt host, we log now:

    not an oVirt host, using non-oVirt host

For any other error, we can degrade to remote transfer, but we want to
warn the user about this, so we log:

    warning: cannot read host id, using non-oVirt host: [Errno 2] No such file or directory: '/etc/vdsm/vdsm.id'

[1] https://github.com/libguestfs/virt-v2v/issues/46#issuecomment-2039604719

RWMJ: Changed "any host" to "non-oVirt host" to make it clearer.