convert: windows: Allow increasing verboseness of driver installation
To try to chase an issue with installing drivers on 32 bit Windows 10,
let's increase the LogLevel to maximum to get maximum logging.  As I
understand it we shouldn't leave this on all the time, so it is
commented out by default.

This setting is explained further here:
https://learn.microsoft.com/en-us/windows-hardware/drivers/install/setting-setupapi-logging-levels

Thanks: Vadim Rozenfeld
Related: https://issues.redhat.com/browse/RHEL-53990