lib, input: Model UEFI secureboot property in metadata
This models a UEFI NVRAM variable which controls whether or not EFI
binary signatures are verified, which is an initial step in the long
chain of operations commonly known as "secure boot".  I added plenty
of documentation explaining what this really means as it is
non-obvious.

Thanks: Daniel Berrange, Andrea Bolognani