in-place: Add new -O option to write inspector XML
When using virt-v2v-in-place there is no easy way to get the
post-conversion metadata, such as the operating system and firmware
that virt-v2v detected inside the guest.

This commit adds new, optional '-O output.xml' to write this
information out to a file.  The format is identical to
virt-v2v-inspector (and roughly a superset of virt-inspector).

Fixes: https://issues.redhat.com/browse/RHEL-58032
Thanks: Martin Necas