tests: Rename all test-v2v-* to test-*
When virt-v2v was part of libguestfs, before it was a separate
component, we prefixed all test names with 'test-v2v-*' in order to
distinguish them from tests for other components.  Since virt-v2v is
now separate, we no longer need to do this.  Thus we can remove the
'-v2v-' part of every test name, which reduces typing and makes the
names nicer.

Note that we cannot remove the 'test-' prefix.  This is because
automake reserves an unspecified list of words (eg. "compile") and
tests using those names will fail.  Keeping the 'test-' prefix avoids
this.