tests: Skip tests if ntfs3g feature not available
If ntfs-3g is not available when libguestfs was built then we cannot
build the Windows phony image, and many tests fail.  (This is the case
if you self-build libguestfs on RHEL for example).  What happens is
that our test suite builds a zero-sized windows.img.  The correct way
to run the test is to require that windows.img has size > 0 (ie.
test -s, not test -f).