v2v: Use nbdcopy --blkhash in verbose mode
In verbose mode, and if nbdcopy supports it, add the nbdcopy --blkhash
option.  This will compute and print a hash of the disk as we are
copying it (more precisely, a hash of the source disk after conversion
changes have been made).  This will allow users to detect any
corruption during or after writing the output.

This feature requires nbdcopy >= 1.23.1

It adds some overhead to the copy, but (a) we're almost always copying
over the network which is slow anyway and (b) this is only done in
verbose mode where there's a lot of overhead from the output anyway.
However keep an eye on this overhead.

Note what is printed is a blkhash, not a regular checksum.  See:
https://gitlab.com/nirs/blkhash/

Fixes: https://issues.redhat.com/browse/RHEL-85508
Fixes: https://issues.redhat.com/browse/RHEL-85512
Fixes: https://issues.redhat.com/browse/RHEL-85514
(cherry picked from commit cffd129d8fd47554255d52ad611d58a30b6b9951)