docs: Clarify --root first documentation
Clarify that we don't necessarily choose the default bootloader option
(since we don't collect that information).  It's just the first in the
list of roots returned by libguestfs.

What is intentionally not documented here is that libguestfs doesn't
necessarily return the roots in any particular order (eg. it's not
sorted alphabetically).  If we fix that in future, we might break how
this option works, so don't document any expectations.

(cherry picked from commit bc936379e20e1aab5f569a577663082411f56dc2)