convert: Split out choosing the root and setting up the inspect struct
This splits out the two separate operations of:

(a) doing libguestfs inspection and choosing the root device to
convert, and

(b) mouting up the filesystems and creating the inspect struct which
is used by conversion.

It's helpful to refer to the basic inspection method in this example:
https://github.com/libguestfs/libguestfs/blob/master/examples/inspect-vm.c

This is neutral refactoring.