convert: Change API for Libosinfo_utils.get_os_by_short_id
Update the common submodule:

  Cole Robinson (1):
      mlcustomize: disable `--inject-virtio-win osinfo`

  Richard W.M. Jones (1):
      mltools: Fix memory leak in OCaml binding of libosinfo

The mlcustomize change is not relevant to virt-v2v since we don't
expose the --inject-virtio-win option.

The commit which fixes the memory leak in libosinfo bindings changes
the API of Libosinfo_utils.get_os_by_short_id, requiring a small
change to convert/convert_linux.ml.