Add new virt-v2v-open tool
This can be used to open the virt-v2v input method and run an almost
arbitrary program on it.  For example you can run virt-inspector to
inspect the input (eg. to find root devices and other information for
the virt-v2v --root option), or guestfish.

Currently this only opens the input read-only, although read-write
access might be added in future.

Suggested-by: Martin Necas
Fixes: https://issues.redhat.com/browse/RHEL-88985