output: Optimize output to raw format file to use single nbdkit instance
In the case where we are writing to multiple raw format, uncompressed
local output files, we can use a single nbdkit instance, and choose
which file we are writing to using the NBD export name.

The nbdkit-file-plugin 'dir' parameter requires nbdkit >= 1.22
(released August 2020).  But actually we already require a later
version of nbdkit, so we don't have to check this.