input: vddk: Use single nbdkit-vddk-plugin instance with exports
nbdkit >= 1.43.8 VDDK plugin supports a new 'export' parameter which
allows the client (virt-v2v) to select which file on the VMware server
to serve.  This allows us to run a single nbdkit instance (if nbdkit
has this feature), which is potentially more efficient for guests with
multiple disks.

This also requires nbdkit-cow-filter to be export-safe.  This change
was made in the same version of nbdkit, so checking for the new
'export' parameter is sufficient.